Fix integration test #40
